now AOK compliant!
For now, and to not risk interferance with other stuff, I added it under
its current name pkg_group
I originally use this in jaclu/spd, so therefore in order to use the AOK
groups param -a has to be given currently to use yours.

So some questions:

A) Is it your preference to keep the bloat group definitions in
/usr/local/etc/AOK_VARS to support other tools? If so then they should stay,
if not it could be put inside pkg_group.

B) If you wanted to keep the groups in its current location, is it ok
to just change the bloat script to run pkg_group -a?
this way I can keep the main code of pkg_group identical in AOK / spd
potentially only changing the group definitions. This would make it much
easier for me to keep updates of pkg_group in sync.

C) If bloat is chaining pkg_group, is maintaining the current syntax for bloat
a must (the group names not matching actual group names) - can certainly be
done, but if it was only used for human consumption, it doesn't serve any
purpose, since that interface, well was sub-optimal :) If it is used from
scripts, then it might make sense to ensure it can be used with the same
notation as before.

D) currently I am removing BLOAT and _APKS from the group names,
and lower casing them, to make it more visually palatable :)
If other tools depend on the current group names when calling bloat
this is super easy to undo, just let me know.