aok rewrite Greatly simplified the script and removd the redundant two step syntax -p login disable Now its just: -l disable
I removed the -v | --verbose option from the help text,
since xset rarely if ever gives much help with a reasonably clean
scipt, but it is still in there so can be used, if something
depends on it. Let me know if you want it mentioned in the help text.